Rename files to be more pythonic
