+++ fault_api.py
